# [MRG] unload data on iteration over SBT leaves